Add support for Cloud Quota Management API
Requests for work with `limits` and `quotas` are now sent to
Cloud Quota Management API instead of Cloud Management API, as
it was before.

Commands for working with limits and quotas have been changed.
Now the commands look like this:
selvpc quota set --region VALUE [--zone VALUE] --resource VALUE --value VALUE <project_id>
selvpc quota show --region VALUE <project_id>
selvpc limit show --region VALUE <project_id>

Cloud Quota Management API is a region service that works with
X-Auth-Token, opposite the Cloud Management API that works with
X-Token and is only in one region.

RegionalHTTPClient is a Proxy of the HTTPClient class, it resolves
the URL of the service in the requested region and adds X-Auth-Token to
the headers.

The functionality for issuing and caching the X-Auth-Token, as
well as resolving the service URL is encapsulated in
the IdentityManager class.

OPENSTACK-12283